chore(subgraph): file rename for clarity
