+++ crates/subgraph/src/lib.rs
@@ -1,4 +1,4 @@
-mod client;
+mod orderbook_client;
 mod csv;
 mod cynic_client;
 mod pagination;
@@ -8,6 +8,6 @@ pub mod utils;
 #[cynic::schema("orderbook")]
 pub mod schema {}
 
-pub use client::{OrderbookSubgraphClient, OrderbookSubgraphClientError};
+pub use orderbook_client::{OrderbookSubgraphClient, OrderbookSubgraphClientError};
 pub use csv::{TryIntoCsv, TryIntoCsvError};
 pub use pagination::{PageQueryClient, PaginationArgs};
